merge-ort: restart merge with cached renames to avoid process entry cost
One of the interesting things about the optimization to do trivial
directory resolutions, is that it often benefits rebases or cherry-picks
and in particular will benefit commits other than the first one in the
series.  The reason for this comes from our previous optimization to
cache known renames and avoid redetecting them.  When we already have
all the relevant renames cached and there are no either no additional
unmatched source pairs or no additional unmatched destination pairs,
then we know precisely which trivial directory resolution candidates we
have to recurse into -- those containing a source or destination path
for one of our renames.  Also, since only a small subset of the
potential renames were cached due to most of the renames being
irrelevant (as noted in another earlier optimization), this is only a
small subset of the directories we need to recurse into (and when we
recurse into them, we only need to keep recursing into subdirectories
that lead us to one of the rename sources or targets; sibling
directories can be trivially resolved).

This has left us with some great speedups, ony preventing us from using
the trivial directory resolution on commits where there is at least one
relevant unmatched (uncached) rename source and one unmatched rename
destination.  However, staring at this for a little while brought us to
another realization: process_entries() is ten to twenty percent more
expensive than collect_merge_info().  So, in our usual merge of running:
   collect_merge_info()
   detect_and_process_renames()
   process_entries()
after detect_and_process_renames() we know that all the directories that
didn't turn out to have renames in them could have been handled a lot
faster by process_entries() if we hadn't recursed into them.  It would
be nice if we could do things over using the information we gained.  So,
we do that:
   collect_merge_info()
   detect_and_process_renames() --> caching all the renames
   collect_merge_info() --> again, but with more trivial dir resolves
   detect_and_process_renames() --> again, but no-op due to cache
   process_entries()
And when we hit process_entries() now it's faster.

However, the above does mean we run collect_merge_info() again which
adds to the running time.  The second one will be cheaper if we were
able to do some trivial directory resolves that we couldn't the first
time around, and it'll make process_entries() cheaper than it would have
been without the redo.  So we use a simple heuristic about how many
paths we might avoid recursing into to guide whether to go this route.
This heuristic uses a "wanted_factor" that I have virtually no actual
real world data for, just some back-of-the-envelope guesses plus picking
a nice round number out of thin air.  It could be that tweaking this
number to make it either higher or lower improves the optimization.

Also, note that while the diffstat looks kind of long (nearly 100
lines), more than half of it is in two comments explaining how things
work.

For the testcases mentioned in commit 557ac0350d ("merge-ort: begin
performance work; instrument with trace2_region_* calls", 2020-10-28),
this change improves the performance as follows:

                            Before                  After
    no-renames:      205.7  ms ±  2.3  ms   206.0  ms ±  2.9  ms
    mega-renames:      2.182 s ±  0.027 s     1.450 s ±  0.021 s
    just-one-mega:   479.5  ms ±  5.1  ms   361.6  ms ±  3.1  ms

Signed-off-by: Elijah Newren <newren@gmail.com>